Move letter and board into separate class files
